+++ src/test/java/org/cicirello/ds/FibonacciHeapDoubleTests.java
@@ -82,6 +82,74 @@ public class FibonacciHeapDoubleTests {
 	}
 	
 	@Test
+	public void testRemoveAll() {
+		String[] elements = {"A", "B", "C", "D"};
+		double[] priorities = { 8, 6, 4, 2 };
+		final FibonacciHeapDouble<String> pq = FibonacciHeapDouble.createMinHeap();
+		ArrayList<PriorityQueueNode.Double<String>> list = new ArrayList<PriorityQueueNode.Double<String>>();
+		for (int i = 0; i < elements.length; i++) {
+			list.add(new PriorityQueueNode.Double<String>(elements[i], priorities[i]));
+		}
+		assertTrue(pq.addAll(list));
+		assertEquals(elements.length, pq.size());
+		assertTrue(pq.removeAll(list));
+		assertEquals(0, pq.size());
+		for (String e : elements) {
+			assertFalse(pq.contains(e));
+		}
+		assertTrue(pq.addAll(list));
+		assertEquals(elements.length, pq.size());
+		list.remove(list.size()-1);
+		assertTrue(pq.removeAll(list));
+		assertEquals(1, pq.size());
+		for (int i = 0; i < elements.length-1; i++) {
+			String e = elements[i];
+			assertFalse(pq.contains(e));
+		}
+		assertTrue(pq.contains(elements[elements.length-1]));
+		
+		list.clear();
+		ArrayList<String> list2 = new ArrayList<String>();
+		ArrayList<String> list3 = new ArrayList<String>();
+		for (int i = 0; i < elements.length; i++) {
+			list2.add(elements[i]);
+			list3.add(elements[i]);
+			list.add(new PriorityQueueNode.Double<String>(elements[i], priorities[i]));
+		}
+		pq.clear();
+		assertTrue(pq.addAll(list));
+		assertEquals(elements.length, pq.size());
+		assertTrue(pq.removeAll(list2));
+		assertEquals(0, pq.size());
+		assertTrue(pq.addAll(list));
+		assertEquals(elements.length, pq.size());
+		list2.remove(list.size()-1);
+		assertTrue(pq.removeAll(list2));
+		assertEquals(1, pq.size());
+		assertFalse(pq.removeAll(list2));
+		assertEquals(1, pq.size());
+		assertTrue(pq.contains(elements[elements.length-1]));
+		for (int i = 0; i < elements.length-1; i++) {
+			String e = elements[i];
+			assertFalse(pq.contains(e));
+		}
+		
+		pq.clear();
+		assertTrue(pq.addAll(list));
+		assertEquals(elements.length, pq.size());
+		list3.remove(0);
+		assertTrue(pq.removeAll(list3));
+		assertEquals(1, pq.size());
+		assertFalse(pq.removeAll(list3));
+		assertEquals(1, pq.size());
+		assertTrue(pq.contains(elements[0]));
+		for (int i = 1; i < elements.length; i++) {
+			String e = elements[i];
+			assertFalse(pq.contains(e));
+		}
+	}
+	
+	@Test
 	public void testIterator() {
 		int n = 4;
 		String[] elements = createStrings(n);